Implemented generalized  vector of pairs to pair of vectors format transformation. (aka  homomorphism!) Unlike the Unboxed Vector format, the tuple structure of the values is exposed, so authors of performance sensitive routines can actually exploit having a struct of arrays format.
These VProd   and MVProd types will work with any vector types, including the standard Boxed,Unboxed
and Stroable types.

One use case would be implementing Complex Number vector and matrix routines
that exploit how the Real and Complex parts are used and can be
easily vectorized when the complex number vector is model as two
vectors for the respective Real and Complex parts.